Clean up Google Config
