+++ tests/components/google_assistant/__init__.py
@@ -1,6 +1,32 @@
-
-
 """Tests for the Google Assistant integration."""
+from homeassistant.components.google_assistant import helpers
+
+
+class MockConfig(helpers.AbstractConfig):
+    """Fake config that always exposes everything."""
+
+    def __init__(self, *, secure_devices_pin=None, should_expose=None,
+                 entity_config=None):
+        self._should_expose = should_expose
+        self._secure_devices_pin = secure_devices_pin
+        self._entity_config = entity_config or {}
+
+    @property
+    def secure_devices_pin(self):
+        """Return secure devices pin."""
+        return self._secure_devices_pin
+
+    @property
+    def entity_config(self):
+        """Return secure devices pin."""
+        return self._entity_config
+
+    def should_expose(self, state):
+        """Expose it all."""
+        return self._should_expose is None or self._should_expose(state)
+
+
+BASIC_CONFIG = MockConfig()
 
 DEMO_DEVICES = [{
     'id':
